Fix arrow label position on Chrome
The commit 5b8adb856bd4085c1f4cf934d19ece98aafcadeb messed with the rendering position of the arrow label on Chrome. More specifically, the commit wrapped the `<g>` tag on ArrowLabel with a `<svg>` so jest wouldn't complain of unknown `<g>` element on ArrowLabel test (jest does that because it has no way to know, from the unit test, that on production there's actually another component wrapping it with a `svg` tag). However, this made the sequence diagram itself to now have nested svg elements, and unfortunatelly, browsers don't handle that consistently. On Firefox the svg coordinate system kept working just fine, but on Chrome the svg nesting resulted in changing the position of the arrow label, causing the bug we observed.

This commit fix that issue by removing the svg wrapping from the source code and, instead, adding it to the ArrowLabel test, so jest gets happy, Chromes gets happy and everybody wins.